remove some old, commented code
