+++ axle-core/src/main/scala/axle/stats/ConditionalProbabilityTable.scala
@@ -68,26 +68,6 @@ case class ConditionalProbabilityTable0[A, N: Field: Order](val p: Map[A, N]) {
   val field = Field[N]
   val order = Order[N]
 
-  def map[B](f: A => B): ConditionalProbabilityTable0[B, N] =
-    ConditionalProbabilityTable0[B, N](
-      values
-        .map({ v => f(v) -> probabilityOf(v) })
-        .groupBy(_._1)
-        .mapValues(_.map(_._2).reduce(field.plus)))
-
-  def flatMap[B](f: A => ConditionalProbabilityTable0[B, N]): ConditionalProbabilityTable0[B, N] =
-    ConditionalProbabilityTable0[B, N](
-      values
-        .flatMap(a => {
-          val p = probabilityOf(a)
-          val subDistribution = f(a)
-          subDistribution.values.map(b => {
-            b -> (p * subDistribution.probabilityOf(b))
-          })
-        })
-        .groupBy(_._1)
-        .mapValues(_.map(_._2).reduce(field.plus)))
-
   val bars = p.scanLeft((dummy[A], field.zero))((x, y) => (y._1, x._2 + y._2)).drop(1)
 
   def values: IndexedSeq[A] = p.keys.toVector
